fix(cppcheck): fix: fix cppcheck codecheck errors

Created-by: corleoneliu
Commit-by: corleoneliu
Merged-by: corleoneliu
Description: <!--  Thanks for sending a pull request!  Here are some tips for you:

1) If this is your first time, please read our contributor guidelines: https://gitee.com/mindspore/mindspore/blob/master/CONTRIBUTING.md

2) If you want to contribute your code but don't know who will review and merge, please add label `mindspore-assistant` to the pull request, we will find and do it as soon as possible.
-->

**What type of PR is this?**
<!-- 
Choose one label from `bug`, `task`, `feature` and `refactor`, and replace `<label>` below the comment block. 

If this pr is not only bugfix/task/feature and also a refactor, you can append `/kind refactor` label after `/kind bug`, `/kind task` and `/kind feature`.
-->
/kind bug


**What does this PR do / why do we need it**:
Clean code:
    - G.CNS.02: replace 27 magic numbers with named constexpr
    - G.PRE.01: replace 4 macro constants with constexpr
    - G.EXP.25: add static_cast for 179 signed/unsigned mismatches
    - G.CMT.02: move 43 inline comments above code
    - G.EXP.17: add static_cast<int64_t> for 2 lossy conversions
    - G.STD.06: replace atoi with strtol
    - R0914: refactor 3 Python functions to reduce local variables
    - G.FMT.05: fix 4 Python import placements
    - G.FNM.07: handle 1 unchecked return value
    - G.CLS.01: add {} default initialization to 5 member variables
    - G.EXP.23: replace bool |= with bool = bool || (13 items)
    - G.EXP.23: cast dlopen/ACL_PROF flags to unsigned (5 items)
    - G.EXP.21: add bounds checks and safe ceiling division (5 items)
    - nullPointer: add explicit null checks for tryGetConfig/getNewShape (23 items)
    - unsafe_function: replace memcpy with std::copy_n
    - G.RES.01: add array bounds check
    - G.RES.06: use explicit lambda captures
    - G.CTL.03: add default case to switch
    - G.FIL.02: add empty path validation (6 items)
    - G.ERR.05: add noexcept to destructor
    - G.EXP.18: refactor switch to unordered_map in GPUKernelToPTX


**Which issue(s) this PR fixes**:
<!-- 
*Automatically closes linked issue when PR is merged.
Usage: `Fixes #<issue number>`, or `Fixes (paste link of issue)`.
-->
Fixes #


**Code review checklist [[illustration]](https://gitee.com/mindspore/community/blob/master/security/code_review_checklist_mechanism.md)**:

- [ ] whether to verify the function's return value (It is forbidden to use void to mask the return values of security functions and self-developed functions. C++ STL functions can be masked if there is no problem)
- [ ] Whether to comply with ***SOLID principle / Demeter's law***
- [ ] Whether there is UT test case && the test case is a valid (if there is no test case, please explain the reason)
- [ ] Whether the API change is involved
- [ ] Whether official document modification is involved

<!-- **Special notes for your reviewers**: -->
<!-- - [ ] Whether it causes forward compatibility failure -->
<!-- - [ ] Whether the dependent third-party library change is involved -->


See merge request: mindspore/akg!2318

+++ akg-mlir/lib/Dialect/Affine/Transforms/MergeFusionOp.cpp
@@ -300,7 +300,7 @@ void MergeFusionOp::runOnOperation() {
       }
     }
 
-    for (int i = num - 1; i >= 0; --i) {
+    for (int i = static_cast<int>(num) - 1; i >= 0; --i) {
       Operation *curOp = band[(unsigned int)i];
       if (mergeSpecificOp) {
         getFusionOpBetweenOp(firstOp->getBlock(), dyn_cast<affine::AffineForOp>(curOp));
